+++ CHANGELOG.md
@@ -1,4 +1,6 @@
+- Add folder filtering context menu
 - Make multiline TODO highlights work properly
+- Add option to disable hiding the tree when it is empty
 
 # v0.0.136 - 2019-06-28
 - Add option to highlight line to full width of the editor